requirements-uv.txt
imageio
numpy<2
scikit-image
scipy
tqdm
typing_extensions
matplotlib
oauthlib
packaging
tensorboard
torchmetrics
lpips
av
imageio[ffmpeg]
